requirements.txt
absl-py==2.1.0
asgiref==3.7.2
astroid==3.0.2
asttokens==2.4.1
astunparse==1.6.3
attrs==23.2.0
autopep8==1.5.5
bleach==6.1.0
blinker==1.7.0
cachetools==5.3.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
crcmod==1.7
cryptography==42.0.2
dbus-python==1.2.16
decorator==5.1.1
Deprecated==1.2.14
dill==0.3.8
distlib==0.3.8
Django==4.2.9
dnspython==2.5.0
exceptiongroup==1.2.0
executing==2.0.1
filelock==3.13.1
flake8==3.8.4
flask==3.0.1
Flask-PyMongo==2.3.0
flatbuffers==23.5.26
gast==0.5.4
gce-v2v @ https://storage.googleapis.com/velostrata-release/gce-v2v/gce-v2v.tar.gz
google-api-core==2.16.2
google-api-python-client==2.116.0
google-auth==2.27.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-cloud-appengine-logging==1.4.0
google-cloud-audit-log==0.2.5
google-cloud-bigquery==3.17.1
google-cloud-core==2.4.1
google-cloud-datastore==2.19.0
google-cloud-language==2.13.0
google-cloud-logging==3.9.0
google-cloud-spanner==3.42.0
google-cloud-storage==2.14.0
google-cloud-translate==3.15.0
google-cloud-videointelligence==2.13.0
google-cloud-vision==3.6.0
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
grpc-google-iam-v1==0.13.0
grpc-interceptor==0.15.4
grpcio==1.60.1
grpcio-status==1.60.1
h5py==3.10.0
httplib2==0.22.0
idna==3.6
importlib-metadata==7.0.1
ipython==8.18.1
isort==5.13.2
itsdangerous==2.1.2
jaraco.classes==3.3.0
jedi==0.17.0
jeepney==0.8.0
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
keras==2.15.0
keyring==24.3.0
keyrings.google-artifactregistry-auth==1.1.2
lazy-object-proxy==1.6.0
libclang==16.0.6
Markdown==3.5.2
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
mccabe==0.6.1
mercurial==5.6.1
ml-dtypes==0.2.0
more-itertools==10.2.0
mypy==1.8.0
mypy-extensions==1.0.0
numpy==1.26.3
oauth2client==4.1.3
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.2
parso==0.8.3
pexpect==4.9.0
platformdirs==4.2.0
pluggy==1.4.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.23.4
psycopg2==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycodestyle==2.6.0
pycparser==2.21
pycurl==7.43.0.6
pydocstyle==6.3.0
pyflakes==2.2.0
pygments==2.17.2
PyGObject==3.38.0
pylint==3.0.3
pyls-mypy==0.1.8
pymongo==4.6.1
pyparsing==3.1.1
python-apt==2.2.1
python-dateutil==2.8.2
python-jsonrpc-server==0.4.0
python-language-server==0.35.1
pytoolconfig==1.3.1
PyYAML==6.0.1
referencing==0.33.0
requests==2.31.0
requests-oauthlib==1.3.1
rope==1.12.0
rpds-py==0.17.1
rsa==4.9
SecretStorage==3.3.3
six==1.16.0
snowballstemmer==2.2.0
sqlparse==0.4.4
stack-data==0.6.3
supervisor==4.2.2
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0.post1
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.35.0
termcolor==2.4.0
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.3
traitlets==5.14.1
typing-extensions==4.9.0
ujson==5.9.0
uritemplate==4.1.1
urllib3==2.2.0
virtualenv==20.25.0
wcwidth==0.2.13
webencodings==0.5.1
werkzeug==3.0.1
wrapt==1.14.1
yapf==0.40.2
zipp==3.17.0
